+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dysonic/dy-cli-core
 
+## 1.0.1
+
+### Patch Changes
+
+- Prevent prerelease publishing from assigning `latest` to packages that do not already have a stable release.
+
 ## 1.0.0
 
 ### Major Changes